Redo the preprocessor
The previous implementation was a mess mainly because I tried to find
algorithmic improvements to the execution algorithm to reduce copying of
tokens. This proved unmaintainable, and now I rewrote it in the image of
the standard's description. The new code is much cleaner and already
works on the cpplib tests; only metalang99 still evades it, for now.